+++ ohkami/src/layer1_req_res/response/headers.rs
@@ -1,114 +1,130 @@
+/* ref: https://developer.mozilla.org/ja/docs/Web/HTTP/Headers */
 #![allow(non_snake_case)]
 
-use crate::layer0_lib::{now, AsStr, ContentType};
+use std::{collections::HashMap};
 
-/// response headers other than `Content-*`
-pub struct ResponseHeaders(String);
 
-impl ResponseHeaders {
-    #[inline(always)] pub(crate) fn new() -> Self {
-        Self(String::from("\
-            Connection: Keep-Alive\r\n\
-            Keep-Alive: timeout=5\r\n\
-        "))
-    }
+struct Header(Option<&'static str>);
 
-    #[inline(always)] pub(crate) fn as_str(&self) -> &str {
-        &self.0
-    }
+pub trait HeaderValue {
+    fn into_header_value(self) -> Option<&'static str>;
 }
+impl HeaderValue for &'static str {fn into_header_value(self) -> Option<&'static str> {Some(self)}}
+impl HeaderValue for Option<&'static str> {fn into_header_value(self) -> Option<&'static str> {self}}
 
-impl ResponseHeaders {
-    #[inline] pub fn append(&mut self, key: &str, value: impl AsStr) {
-        self.0.push_str(key);
-        self.0.push(':');
-        self.0.push(' ');
-        self.0.push_str(value.as_str());
-        self.0.push('\r');
-        self.0.push('\n');
-    }
+macro_rules! ResponseHeaders {
+    ($(
+        $group:ident {
+            $( $key:literal $name:ident( $arg:ident ), )*
+        }
+    )*) => {
+        pub struct ResponseHeaders {
+            $( $group: bool, )*
+            $($( $name: Header, )*)*
+            custom: HashMap<&'static str, &'static str>,
+        }
 
-    pub fn clear(&mut self, key: &str) {
-        let mut pos = 0;
-        for (k, v) in self.0.lines().map(|kv| unsafe {kv.split_once(':').unwrap_unchecked()}) {
-            if k != key {
-                pos += k.len() + 1/* ':' */ + v/* with ' ' */.len() + 2/* "\r\n" */
-            } else {
-                let prev_end = pos - 2/* "\r\n" */;
-                let mut to = prev_end + k.len() + 1/*':'*/ + v/*with ' '*/.len() - 1;
-                if v.ends_with('\n') {to += 2/* "\r\n" */};
-                self.0.replace_range(prev_end..to, "");
-                return
+        impl ResponseHeaders {
+            $($(
+                pub fn $name(&mut self, $arg: impl HeaderValue) -> &mut Self {
+                    self.$group = true;
+                    self.$name.0 = $arg.into_header_value();
+                    self
+                }
+            )*)*
+
+            pub(crate) fn new() -> Self {
+                Self {
+                    $( $group: false, )*
+                    $($( $name: Header(None), )*)*
+                    custom: HashMap::new()
+                }
+            }
+
+            pub(crate) fn to_string(&self) -> String {
+                let mut h = format!("\
+                    Connection: Keep-Alive\r\n\
+                    Keep-Alive: timout=5\r\n\
+                    Date: {}\r\n\
+                ", crate::layer0_lib::now());
+
+                $(
+                    if self.$group {
+                        $(
+                            if let Some(value) = self.$name.0 {
+                                h.push_str($key);
+                                h.push_str(value);
+                                h.push('\r'); h.push('\n');
+                            }
+                        )*
+                    }
+                )*
+
+                h.push('\r'); h.push('\n'); h
             }
         }
+    };
+} ResponseHeaders! {
+    auth_cookie_security_context {
+        // authentication
+        "WWW-Authenticate: "                 WWWAuthenticate(challenge),
+        // cookie
+        "Set-Cookie: "                       SetCookie(cookie_and_directives),
+        // security
+        "X-Frame-Options: "                  XFrameOptions(DENY_or_SAMEORIGIN),
+        // response context
+        "Allow: "                            Allow(methods),
+        "Server: "                           Server(product),
     }
 
-    pub fn set(&mut self, key: &str, value: impl AsStr) {
-        let mut pos = 0;
-        let mut v_range = None;
+    cache_proxy_redirect_others {
+        // cache
+        "Age: "                              Age(delta_seconds),
+        "Cache-Control: "                    CacheControl(cache_control),
+        "Expires: "                          Expires(http_date),
+        // proxy
+        "Via: "                              Via(via),
+        // redirect
+        "Location: "                         Location(url),
+        // others
+        "Alt-Srv: "                          AltSvc(alternative_services),
+    }
 
-        for (k, v) in self.0.lines().map(|kv| unsafe {kv.split_once(':').unwrap_unchecked()}) {
-            if k != key {
-                pos += k.len() + 1/* ':' */ + v/* with ' ' */.len() + 2/* "\r\n" */
-            } else {
-                let value_start = pos + k.len() + 2/* ": " */;
-                v_range.replace(value_start..(value_start+v.len()-1));
-                break
-            }
-        }
+    conditions {
+        "Last-Modified: "                    LastModified(http_date),
+        "E-tag: "                            Etag(identical_string),
+        "If-Match: "                         IfMatch(etag_values),
+        "If-None-Match: "                    IfNoneMatch(etag_values),
+        "If-Modified-Since: "                IfModifiedSince(http_date),
+        "If-Unmodified-Since: "              IfUnmodifiedSince(http_date),
+        "Vary: "                             Vary(header_names),
+    }
 
-        match v_range {
-            Some(range) => self.0.replace_range(range, value.as_str()),
-            None => self.append(key, value)
-        }
+    cors {
+        "Access-Control-Allow-Origin: "      AccessControlAllowOrigin(origin),
+        "Access-Control-Allow-Credentials: " AccessControlAllowCredentials(true_if_needed),
+        "Access-Control-Allow-Headers: "     AccessControlAllowHeaders(headers),
+        "Access-Control-Allow-Methods: "     AccessControlAllowMethods(methods),
+        "Access-Control-Expose-Headers: : "  AccessControlExposeHeaders(headers),
+        "Access-Control-Max-Age: : "         AccessControlMaxAge(delta_seconds),
+    }
+
+    message_body_and_encoding {
+        // message body
+        "Content-Encoding: "                 ContentEncoding(algorithm),
+        "Content-Language: "                 ContentLanguage(language_tag),
+        "Content-Location: "                 ContentLocation(url),
+        // transfer encoding
+        "Tranfer-Encoding: "                 TransferEncoding(chunked_compress_deflate_gzip_identity),
+        "Trailer: "                          Trailer(header_names),
     }
 }
 
-
-
-
-#[cfg(test)] #[test] fn check_set_header() {
-    let mut h = ResponseHeaders::new();
-    h.set("Keep-Alive", "timeout=10");
-    assert_eq!(h.as_str(),
-"Connection: Keep-Alive\r
-Keep-Alive: timeout=10"
-    );
-
-    let mut h = ResponseHeaders::new();
-    h.append("Content-Language", "en-US");
-    h.append("Content-Type", "application/json");
-    h.set("Content-Type", "text/html");
-    assert_eq!(h.as_str(),
-"Connection: Keep-Alive\r
-Keep-Alive: timeout=5\r
-Content-Language: en-US\r
-Content-Type: text/html"
-    );
-}
-
-
-#[cfg(test)] #[test] fn check_clear_header() {
-    let mut h = ResponseHeaders::new();
-    h.clear("Connection");
-    assert_eq!(h.as_str(),
-"Keep-Alive: timeout=10"
-    );
-
-    let mut h = ResponseHeaders::new();
-    h.append("Content-Language", "en-US");
-    h.append("Content-Type", "application/json");
-
-    h.clear("Content-Language");
-    assert_eq!(h.as_str(),
-"Connection: Keep-Alive\r
-Keep-Alive: timeout=5\r
-Content-Type: text/html"
-    );
-
-    h.clear("Content-Type");
-    assert_eq!(h.as_str(),
-"Connection: Keep-Alive\r
-Keep-Alive: timeout=5"
-    );
+impl ResponseHeaders {
+    pub fn costom(&mut self, key: &'static str, value: impl HeaderValue) {
+        match value.into_header_value() {
+            Some(value) => {self.custom.insert(key, value);}
+            None        => {self.custom.remove(key);}
+        }
+    }
 }
